requirements-dev.txt
# ~~ Generated by projen. To modify, edit .projenrc.py and run "npx projen".
hallcor.pulumi-projen-project-types
numpy
projen>=0.91.0
pyfakefs
pytest==7.4.3
